move UserControler to API folder
