+++ routes/api.php
@@ -16,8 +16,8 @@ use Illuminate\Http\Request;
 Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
-Route::post('auth/login', 'UserController@login');
-Route::post('auth/register', 'UserController@register');
+Route::post('auth/login', 'API\UserController@login');
+Route::post('auth/register', 'API\UserController@register');
 Route::group(['middleware' => 'jwt.auth'], function () {
-    Route::get('user-info', 'UserController@getUserInfo');
+    Route::get('user-info', 'API\ProfileController@show');
 });
